split util functions
